# docs/api_spec.yaml
openapi: 3.0.0
info:
  title: Face Recognition API
  version: 1.0.0
paths:
...
  /api/users:
    post:
      summary: Create a new user account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: User created successfully
  /api/users/me:
    get:
      summary: Get current user information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      summary: Update current user information
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: User updated successfully
    delete:
      summary: Delete current user account
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User deleted successfully
...
  /api/subscriptions:
    post:
      summary: Create a new subscription
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionCreate'
      responses:
        '200':
          description: Subscription created successfully
  /api/subscriptions/me:
    get:
      summary: Get current user's active subscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
    put:
      summary: Cancel current user's subscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription cancelled successfully
  /api/subscriptions/history:
    get:
      summary: Get subscription history for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionResponse'
...
  /api/usage/current:
    get:
      summary: Get current month's usage for authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current usage statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
  /api/usage/limits:
    get:
      summary: Get usage limits for current user's plan
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usage limits
...
  /api/support/tickets:
    get:
      summary: Get all support tickets for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of support tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupportTicketResponse'
    post:
      summary: Create a new support ticket
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportTicketCreate'
      responses:
        '200':
          description: Support ticket created successfully
  /api/support/tickets/{ticket_id}:
    get:
      summary: Get a specific support ticket
      security:
        - bearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Support ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportTicketResponse'
    put:
      summary: Update a support ticket
      security:
        - bearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportTicketUpdate'
      responses:
        '200':
          description: Support ticket updated successfully
    delete:
      summary: Delete a support ticket
      security:
        - bearerAuth: []
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Support ticket deleted successfully
components:
  schemas:
...
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        subscription_tier:
          type: string
          default: free
    UserResponse:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
        name:
          type: string
        subscription_tier:
          type: string
        created_at:
          type: string
...
    SubscriptionCreate:
      type: object
      properties:
        plan_id:
          type: string
    SubscriptionResponse:
      type: object
      properties:
        subscription_id:
          type: string
        user_id:
          type: string
        plan_id:
          type: string
        status:
          type: string
        created_at:
          type: string
        expires_at:
          type: string
...
    UsageResponse:
      type: object
      properties:
        user_id:
          type: string
        period_start:
          type: string
        period_end:
          type: string
        recognitions_used:
          type: integer
        enrollments_used:
          type: integer
        recognitions_limit:
          type: integer
        enrollments_limit:
          type: integer
...
    SupportTicketCreate:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
          default: medium
    SupportTicketUpdate:
      type: object
      properties:
        description:
          type: string
        priority:
          type: string
    SupportTicketResponse:
      type: object
      properties:
        ticket_id:
          type: string
        user_id:
          type: string
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT